refactor: refactor structure of header
