+++ src/components/organisms/Header/Header.tsx
@@ -1,17 +1,19 @@
+import { useAppSelector } from "../../../app/reducers/reduxHooks";
 import { useAuth } from "../../../hooks/useAuth";
-import { useModal } from "../../../hooks/useModal";
 import { useNav } from "../../../hooks/useNav";
 import { Btn } from "../../atoms/Btn/Btn";
 import { Logo } from "../../atoms/Logo/Logo";
 import { HeaderList } from "../../molecules/HeaderList/HeaderList";
 
 import S from "./Header.module.scss";
-import { useHeader } from "./useHeader";
+const links = [
+  { to: "/", text: "Главная" },
+  { to: "/tickets", text: "История" },
+];
 
 export const Header = () => {
-  const { links, isAuth, isMainPage } = useHeader();
+  const { isAuth } = useAppSelector((state) => state.auth);
   const { goToLogin } = useNav();
-  const { openCreateModal } = useModal();
   const { toLogOut } = useAuth();
 
   return (
@@ -19,17 +21,13 @@ export const Header = () => {
       <div className={S.logo}>
         <Logo />
       </div>
-      {isAuth && <HeaderList elements={links} />}
-
-      {isMainPage && (
-        <Btn color="grey" type="button" onClick={openCreateModal}>
-          Новое обращение
-        </Btn>
-      )}
       {isAuth ? (
-        <Btn color="grey" type="button" onClick={toLogOut}>
-          Выход
-        </Btn>
+        <>
+          <HeaderList elements={links} />{" "}
+          <Btn color="grey" type="button" onClick={toLogOut}>
+            Выход
+          </Btn>
+        </>
       ) : (
         <Btn color="grey" type="button" onClick={goToLogin}>
           Войти
